reorganize structure of the app #38
