+++ src/controllers/homeController.js
@@ -1,0 +1,19 @@
+const getData = require('../queries/get.js');
+
+const homeRouter = (req, res, next) => {
+  getData.getRecipes()
+    .then((recipeArray) => {
+      const data = {
+        title: 'Recipes',
+        recipes: recipeArray,
+      };
+      res.render('home', data);
+    })
+    .catch((getDataError) => {
+      next(getDataError);
+    });
+};
+
+module.exports = {
+  homeRouter,
+};
